requirements-test-2.10.7.txt
ansible-base==2.10.17
ansible-lint==5.4.0
molecule
molecule-docker
pytest-testinfra
